+++ app/controllers/residents_controller.rb
@@ -14,6 +14,11 @@ class ResidentsController < ApplicationController
   # GET /residents/new
   def new
     @resident = Resident.new
+    if resident_params
+      property_id = resident_params[:residencies_attributes][:property_id]
+      @property = Property.find(property_id)
+      @resident.residencies.build(property: @property)
+    end
   end
 
   # GET /residents/1/edit